backend/pyproject.toml
[project]
name = "wheat-from-chaff-backend"
version = "0.1.0"
description = "FastAPI backend for the candidate filtering system."
requires-python = ">=3.8"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "sqlalchemy",
    "pydantic",
    "python-multipart",
    "httpx",
    "omegaconf",
    "azure-storage-blob",
    "azure-identity",
    "pdf2image",
    "Pillow",
]

[project.scripts]
wfc-serve = "backend.app.main:main"
wfc-db-seed = "backend.scripts.seed_db:seed_database"

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta" 